pyproject.toml
[project]
name = "accountmanager"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "customtkinter (>=5.2.2,<6.0.0)",
    "sc2reader (>=1.8.0,<2.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "streamlit (>=1.55.0,<2.0.0)",
    "pandas (<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
